code/requirements.txt

python-miio
